% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbstan.R
\name{tmbstan}
\alias{tmbstan}
\title{Draw MCMC samples from a TMB model object using STAN}
\usage{
tmbstan(obj, ..., lower = numeric(0), upper = numeric(0), laplace = FALSE,
  silent = TRUE, debug = FALSE)
}
\arguments{
\item{obj}{TMB model object.}

\item{...}{Passed to \code{rstan::sampling} with some modifications - see details.}

\item{lower}{Vector of lower parameter bounds.}

\item{upper}{Vector of upper parameter bounds.}

\item{laplace}{Apply the Laplace approximation to \code{random} subset of parameters ? The default disables the Laplace approximation.}

\item{silent}{Be silent during samling ?}

\item{debug}{Should not be used.}
}
\value{
Object of class \code{stanfit}
}
\description{
Draw MCMC samples from a TMB model object using STAN
}
\details{
\code{tmbstan} works for models with or without random effects.

By default a full Bayesian analysis is carried out, i.e. both
parameters and random effects are sampled using MCMC. Models with
random effects will thus have the Laplace approximation disabled. It
is possible to mix the Laplace approximation with MCMC by setting
\code{laplace=TRUE}.
All methods provided by the \code{rstan} package can be applied to a
fitted object. Get a complete list using
\code{methods(class="stanfit")}.

The function arguments \code{...} are passed to \code{rstan}s
fitting function. For a complete list see \code{?rstan::sampling}. A
few notable arguments are:
\itemize{
\item \code{chains} The number of chains.
\item \code{iter}   The number of iterations.
\item \code{init}   Initial values for the sampler.
Differs slightly from \code{rstan}. If unspecified, the best encountered
parameter \code{obj$env$last.par.best} is used (The MLE with corresponding posterior modes if an
optimization has been carried out - otherwise simply \code{obj$par}). We also
allow to pass a single numeric, or a list of numerics with same length
as the number of chains. Numeric vectors should have the same length
as the number of sampled parameters and names are currently ignored. Parameters
that do not follow the previous scheme (e.g. characters) are passed on
to \code{rstan} unchanged. If in doubt, use \code{rstan::get_inits} to inspect the
applied initial values.
\item \code{seed} Random seed.
}
}
\examples{
runExample("simple")
fit <- tmbstan(obj, chains=1)

## The avilable methods are
methods(class="stanfit")

\dontrun{
## Pairs plot
pairs(fit, pars=names(obj$par))
}

## Trace plot
traceplot(fit, pars=names(obj$par), inc_warmup=TRUE)
}

